requirements.txt
requests
beautifulsoup4
Pillow
mysql-connector-python
